+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/implementation/RecoveryServicesBackupClientImpl.java
@@ -84,8 +84,6 @@ import reactor.core.publisher.Mono;
 /** Initializes a new instance of the RecoveryServicesBackupClientImpl type. */
 @ServiceClient(builder = RecoveryServicesBackupClientBuilder.class)
 public final class RecoveryServicesBackupClientImpl implements RecoveryServicesBackupClient {
-    private final ClientLogger logger = new ClientLogger(RecoveryServicesBackupClientImpl.class);
-
     /** The subscription Id. */
     private final String subscriptionId;
 
@@ -904,7 +902,7 @@ public final class RecoveryServicesBackupClientImpl implements RecoveryServicesB
                             managementError = null;
                         }
                     } catch (IOException | RuntimeException ioe) {
-                        logger.logThrowableAsWarning(ioe);
+                        LOGGER.logThrowableAsWarning(ioe);
                     }
                 }
             } else {
@@ -963,4 +961,6 @@ public final class RecoveryServicesBackupClientImpl implements RecoveryServicesB
             return Mono.just(new String(responseBody, charset));
         }
     }
+
+    private static final ClientLogger LOGGER = new ClientLogger(RecoveryServicesBackupClientImpl.class);
 }