+++ sdk/schemaregistry/azure-data-schemaregistry/src/main/java/com/azure/data/schemaregistry/models/SchemaProperties.java
@@ -6,6 +6,7 @@ package com.azure.data.schemaregistry.models;
 import com.azure.core.annotation.Immutable;
 import com.azure.data.schemaregistry.SchemaRegistryAsyncClient;
 import com.azure.data.schemaregistry.SchemaRegistryClient;
+import com.azure.data.schemaregistry.implementation.SchemaRegistryHelper;
 
 /**
  * Stores properties of a schema stored in Schema Registry.
@@ -18,6 +19,17 @@ public final class SchemaProperties {
 
     private final String id;
     private final SchemaFormat format;
+    private final String groupName;
+    private final String name;
+
+    static {
+        SchemaRegistryHelper.setAccessor(new SchemaRegistryHelper.SchemaRegistryModelsAccessor() {
+            @Override
+            public SchemaProperties getSchemaProperties(String id, SchemaFormat format, String groupName, String name) {
+                return new SchemaProperties(id, format, groupName, name);
+            }
+        });
+    }
 
     /**
      * Initializes a new instance.
@@ -26,8 +38,22 @@ public final class SchemaProperties {
      * @param format The type of schema, e.g. avro, json.
      */
     public SchemaProperties(String id, SchemaFormat format) {
+        this(id, format, null, null);
+    }
+
+    /**
+     * Initializes a new instance.
+     *
+     * @param id The schema id.
+     * @param format The type of schema, e.g. avro, json.
+     * @param groupName The schema group for this schema.
+     * @param name The name of the schema.
+     */
+    SchemaProperties(String id, SchemaFormat format, String groupName, String name) {
         this.id = id;
         this.format = format;
+        this.groupName = groupName;
+        this.name = name;
     }
 
     /**
@@ -45,5 +71,23 @@ public final class SchemaProperties {
      */
     public SchemaFormat getFormat() {
         return format;
+    }
+
+    /**
+     * Gets the schema group of this schema.
+     *
+     * @return The schema group of this schema.
+     */
+    public String getGroupName() {
+        return groupName;
+    }
+
+    /**
+     * Gets the name of the schema.
+     *
+     * @return The name of the schema.
+     */
+    public String getName() {
+        return name;
     }
 }