+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/FluxAutoComplete.java
@@ -22,10 +22,11 @@ import static com.azure.messaging.servicebus.implementation.ServiceBusConstants.
  * Flux operator that auto-completes or auto-abandons messages when control is returned successfully.
  */
 final class FluxAutoComplete extends FluxOperator<ServiceBusMessageContext, ServiceBusMessageContext> {
+    private static final ClientLogger LOGGER = new ClientLogger(FluxAutoComplete.class);
+
     private final Semaphore completionLock;
     private final Function<ServiceBusMessageContext, Mono<Void>> onComplete;
     private final Function<ServiceBusMessageContext, Mono<Void>> onAbandon;
-    private final ClientLogger logger = new ClientLogger(FluxAutoComplete.class);
 
     FluxAutoComplete(Flux<? extends ServiceBusMessageContext> upstream, Semaphore completionLock,
                      Function<ServiceBusMessageContext, Mono<Void>> onComplete,
@@ -46,7 +47,7 @@ final class FluxAutoComplete extends FluxOperator<ServiceBusMessageContext, Serv
         Objects.requireNonNull(coreSubscriber, "'coreSubscriber' cannot be null.");
 
         final AutoCompleteSubscriber subscriber =
-            new AutoCompleteSubscriber(coreSubscriber, completionLock, onComplete, onAbandon, logger);
+            new AutoCompleteSubscriber(coreSubscriber, completionLock, onComplete, onAbandon, LOGGER);
 
         source.subscribe(subscriber);
     }